Add wait-for-it.sh, remove entrypoint
This allows us to craft a command that waits for db to be ready on a
port before starting the flyway migrations.  This helps with things like
"docker-compose up", as now we can automatically run the migration when
the db is ready instead of manually monitoring db state and running
another manual command to start the migrations.